+++ CHANGES.md
@@ -16,6 +16,10 @@
     - `RDI` (or `DSI`) uses now `SWIR_1` instead of `SWIR_2` (see [this](https://github.com/awesome-spectral-indices/awesome-spectral-indices/issues/18) issue)
     - `PANI` equivalent is now `BITM` and is normalised ! (/3.)
     - `SBI` is normalized (/2.) to fit with `BIXS` definition
+- **BREAKING CHANGES: Using `pyresample` to geocode Sentinel-3 data**  ([#55](https://github.com/sertit/eoreader/issues/55))
+  - Cleaner: better conversion from swath to grid
+  - Faster: Up to 4 times faster
+  - Allows code refactoring between OLCI and SLSTR
 
 ### Bug Fixes